requirements.txt
# ================================
# Python Dependencies
# Portfolio Assistant Backend
# Groq-powered LLM (Free Tier)
# ================================

# FastAPI Framework
fastapi==0.104.1
uvicorn==0.24.0

# Groq LLM API Client
groq==1.0.0

# Environment Variables
python-dotenv==1.0.0

# CORS Support
# (Included with FastAPI: fastapi.middleware.cors)

# -------------------------------
# Optional: Production Deployment
# -------------------------------
# gunicorn==21.2.0
# python-multipart==0.0.6

# -------------------------------
# Optional: Database / Async (Future)
# -------------------------------
# sqlalchemy==2.0.23
# databases==0.8.0
